+++ tests/Unit/Controller/LoginTest.php
@@ -12,9 +12,9 @@ namespace OxidEsales\GraphQL\Base\Tests\Unit\Controller;
 use Lcobucci\JWT\Parser;
 use Lcobucci\JWT\ValidationData;
 use OxidEsales\GraphQL\Base\Controller\Login;
-use OxidEsales\GraphQL\Base\Service\AuthenticationService;
-use OxidEsales\GraphQL\Base\Service\KeyRegistryInterface;
-use OxidEsales\GraphQL\Base\Service\LegacyServiceInterface;
+use OxidEsales\GraphQL\Base\Service\Authentication;
+use OxidEsales\GraphQL\Base\Service\KeyRegistry;
+use OxidEsales\GraphQL\Base\Service\Legacy;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -31,11 +31,18 @@ class LoginTest extends TestCase
 
     public function setUp(): void
     {
-        $this->keyRegistry = $this->getMockBuilder(KeyRegistryInterface::class)->getMock();
+        $this->keyRegistry = $this->getMockBuilder(KeyRegistry::class)
+                                  ->disableOriginalConstructor()
+                                  ->getMock();
         $this->keyRegistry->method('getSignatureKey')
             ->willReturn('5wi3e0INwNhKe3kqvlH0m4FHYMo6hKef3SzweEjZ8EiPV7I2AC6ASZMpkCaVDTVRg2jbb52aUUXafxXI9/7Cgg==');
-        $this->legacyService         = $this->getMockBuilder(LegacyServiceInterface::class)->getMock();
-        $this->authenticationService = new AuthenticationService($this->keyRegistry, $this->legacyService);
+        $this->legacy         = $this->getMockBuilder(Legacy::class)
+                                     ->disableOriginalConstructor()
+                                     ->getMock();
+        $this->authentication = new Authentication(
+            $this->keyRegistry,
+            $this->legacy
+        );
     }
 
     public function testCreateTokenWithValidCredentials(): void
@@ -47,14 +54,15 @@ class LoginTest extends TestCase
         $user = [
             'username' => 'admin',
             'password' => 'admin',
-            'group'    => LegacyServiceInterface::GROUP_ADMIN,
+            'group'    => Legacy::GROUP_ADMIN,
         ];
 
-        $this->legacyService->method('checkCredentials');
-        $this->legacyService->method('getUserGroup')->willReturn($user['group']);
-        $this->legacyService->method('getShopUrl')->willReturn($shop['url']);
-        $this->legacyService->method('getShopId')->willReturn($shop['id']);
-        $loginController = new Login($this->authenticationService);
+        $this->legacy->method('checkCredentials');
+        $this->legacy->method('getUserGroup')->willReturn($user['group']);
+        $this->legacy->method('getShopUrl')->willReturn($shop['url']);
+        $this->legacy->method('getShopId')->willReturn($shop['id']);
+
+        $loginController = new Login($this->authentication);
 
         $token = (new Parser())->parse($loginController->token($user['username'], $user['password']));